사이드 메뉴 분리 및 드롭박스 추가
